Refactoring Coroutines Homework
