--- Coroutines-Homework-development/app/src/main/res/values/strings.xml
@@ -1,5 +1,0 @@
-<resources>
-    <string name="app_name">Cat Facts </string>
-    <string name="more_facts">More Facts</string>
-    <string name="no_server_response">Не удалось получить ответ от сервера</string>
-</resources>
