--- Coroutines-Homework-development/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Sun Apr 18 11:19:35 MSK 2021
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-7.4.2-bin.zip
