--- app/src/main/java/otus/homework/coroutines/CatsService.kt
@@ -1,10 +1,0 @@
-package otus.homework.coroutines
-
-import retrofit2.Call
-import retrofit2.http.GET
-
-interface CatsService {
-
-    @GET("fact")
-    fun getCatFact() : Call<Fact>
-}
