--- app/src/main/java/otus/homework/coroutines/DiContainer.kt
@@ -1,16 +1,0 @@
-package otus.homework.coroutines
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-class DiContainer {
-
-    private val retrofit by lazy {
-        Retrofit.Builder()
-            .baseUrl("https://catfact.ninja/")
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-    }
-
-    val service by lazy { retrofit.create(CatsService::class.java) }
-}
