+++ app/src/main/java/otus/homework/coroutines/api_services/DiContainer.kt
