+++ app/src/main/java/otus/homework/coroutines/data/Cat.kt
