+++ app/src/main/java/otus/homework/coroutines/data/CatImage.kt
